resto/backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
django==5.2.6
    # via
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
django-cors-headers==4.8.0
    # via backend
django-filter==25.1
    # via backend
djangorestframework==3.16.1
    # via backend
idna==3.10
    # via requests
pillow==11.3.0
    # via backend
requests==2.32.5
    # via backend
sqlparse==0.5.3
    # via django
tzdata==2025.2 ; sys_platform == 'win32'
    # via django
urllib3==2.5.0
    # via requests
